第二支單筆查詢API 完成

+++ src/main/java/com/example/eggcakeshopapi/service/ProductService.java
@@ -9,11 +9,11 @@ public interface ProductService {
     //  查詢Product(所有)
     List<Product> getAllProducts();
     //  查詢Product=>ID
-    Product getProductById(Integer productId) ;
+    Product getProductById(Long productId) ;
     //   新增Product=>ID
     Integer createProduct(ProductRequest productRequest);
     //   更新Product=>ID
-    void updateProduct(Integer productId,ProductRequest productRequest);
+    void updateProduct(Long productId,ProductRequest productRequest);
     //   刪除Product=>ID
-    void deleteProduct(Integer productId);
+    void deleteProduct(Long productId);
 }
